+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to YTM-Synk are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2026-07-10
+
+### Added
+
+- Progress indicator on the active playlist in the Playlists list while syncing
+
+### Changed
+
+- Disable all playlist Sync buttons while any sync is running
+- Disable Sync for playlists that are not selected
+
 ## [1.3.0] - 2026-07-10
 
 ### Added